+++ WebCLI/views/algorithm_details_view.py
@@ -4,7 +4,7 @@ from django.core.exceptions import PermissionDenied
 from ..models import Algorithm, Molecule, Algorithm_version, Metrics
 from django.db.models import F
 from WebCLI.misc.helpers import get_metrics, get_selected_version
-from WebCLI.misc.helpers import get_selected_metrics, to_positive_int_or_none
+from WebCLI.misc.helpers import get_selected_metrics, get_versions, to_positive_int_or_none
 
 
 def get_algorithm_details_view_params(request):
@@ -17,13 +17,6 @@ def get_algorithm_details_view_params(request):
         "metrics_id": metrics_id,
         "molecule_id": molecule_id
     }
-
-
-def get_versions(algorithm):
-    query = Algorithm_version.objects.filter(algorithm_id=algorithm)
-    query = query.annotate(version_number=RawSQL("ROW_NUMBER() OVER(ORDER BY timestamp)", []))
-    query = query.order_by('-timestamp')
-    return query
 
 
 def get_selected_molecule(params, molecules_with_metrics):
